Added checkedRead.
This will allow us to deal with arbitrary types of arguments
(as long as they have matched Read/Show instances).

We run checkedRead inside the interpreter.  It gives us either
Right and then a value we can put in the program, or left
and an error we can raise with a line number.